revert and moved IAudioPlayer into be package
