+++ src/java/be/IAudioPlayer.java
@@ -1,4 +1,4 @@
-package bll;
+package be;
 
 public interface IAudioPlayer
 {
