+++ src/java/bll/LocalAudio.java
@@ -1,10 +1,10 @@
 package bll;
 
+import be.IAudio;
+import be.IAudioPlayer;
 import be.SongModel;
 
 import java.io.File;
-import java.io.FileReader;
-import java.nio.file.FileSystem;
 
 public class LocalAudio implements IAudio, IAudioPlayer
 {
